demo-applications/scrumboard/requirements.txt
Django==1.3
django-tastypie==0.9.11
decorator==3.3.2